etl-backend/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
flask-socketio==5.3.6
python-dotenv==1.0.1
celery==5.4.0
redis==5.0.4
psycopg2-binary==2.9.9
requests==2.32.3
pandas==2.2.3
openpyxl==3.1.5
pdfplumber==0.11.4
python-dateutil==2.9.0.post0
pgvector==0.3.6
pytest==8.3.3
